requirements.txt
lark-oapi>=1.3.0
openai>=1.55.0,<3.0.0
rembg[cpu]>=2.0.50
Pillow>=10.0.0
tenacity>=8.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
PyJWT>=2.8.0
python-docx>=1.0.0
passlib>=1.7.4
python-multipart>=0.0.6
playwright>=1.44.0
